+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Remove error logs
 * Fix bookings query with prepare, that prevented case insensitive searches
 * Prevent URL too long in bookings list search
+* Display success message after changing booking status
 
 ### 1.6.0: June 17th, 2021
 * Remove special chars from tokens